underscore variable handling
