+++ includes/init/init.h
@@ -2,6 +2,5 @@
 # define INIT_H
 
 void		parse_options(int argc, char **argv, t_shell_env *shell_env);
-void		set_default_variables(void);
 
 #endif
