+++ srcs/execution/execute_builtin.c
@@ -15,6 +15,7 @@ t_error_id	execute_builtin(t_simple_command *cmd, size_t lvl)
 		builtin = get_matching_builtin(cmd->argv[0]);
 		if (builtin == NULL)
 			return (NO_SUCH_BUILTIN);
+		set_variable("_", builtin->name, false);
 		print_n_char_fd(' ', (lvl) * 2, 2);
 		dprintf(2, "executing builtin %s\n", builtin->name);
 		if (builtin->builtin == NULL)
